Make YetAnotherGlobal variable, this one for the main LDraw file.  Right now, this is used from CSI, so that it knows where to add lines representing its dimensions.  Global LDraw file will eventually be used by PLI, Page, and anything else that needs to add itself to the file.
Give CSI a resize method, which is called whenever the user changes the global window size (through the setSize method of Instructions (TODO: rename that to resize)).  No more weird workaround needed to retrieve the CSI bounding box now; the box is always in an consistent state, and so can be used by PLI directly, and can be written to the LDraw file easily.

Give Step a file line, which is a reference to that Step's line in the LDraw file array.  Also, store the new CSI lines to the global LDraw file after we initialize the CSI dimensions (through initLayout).  This store shouldn't happen automatically, but only if no CSI line exists after a STEP lin in the loading file.