Add support for loading PLI and CSI commands from a Lic processed LDraw file.  Add a bunch of parsing assistants to handle new commands to LDrawFileFormat.  Move CSI dimension save/load from where it was with the Parts to the LDraw file where it belongs - moved all the now useless code to Snippets.  Be sure to ignore any initSize & initLayout calls if the respective object has already been loaded from file.
Needed to add a new 'nextStepLine' attribute to the Step class.  This is needed because we must list PLI items *after* all the parts in a step.  If we don't do this, the PLI item can't be created properly, becuase it requires a PartOGL.  So, refactor PLI saving so that all PLI stuff is inserted at the end of the current step (ie, right before the next step line).

A few problems: one, loading still seems sluggish, even if initializing everything from file.  Look into that.  Two: sub model CSIs and PLIs are not being created properly.