Render each step dat that we created in the previous commit.  We've now got all the images we'll need for complete instruction book creation.
One problem remains with the images: they need to be the same size as what OpenGL calculated them to be, and they need to be offset by whatever OpenGL offsets them, to get them to render at the same size and in the same spot as OpenGL expects.  Otherwise, placing each image inside an instruction page will be damn near impossible.  Should be easy though, just need to store what image size was successful in rendering from OpenGL, and pass that, along with offset info already stored, to pov image creation.  Coming right up.

Fix povray wrapper so that it (slightly) better handles Pov-Ray paths, at least the default install path for pov 3.6 on win32.  Also, encase the first arg to the povray process in quotes - it helps when the executable has spaces in the name.