Add image size and a bit of relocation to pov image generation.
When fixing pov files, need to look at the part's actual center, not the center of the scene.  Center is conveniently calculated by l3p, needed to track it down and use it.  Done.

Fix part cache reading, by taking into account part size and lack of beginning 'p'.  Though, given that we have the width and height of the actual part, don't really need the image size it was generated at.  Probably can remove all that.

Some problems remain.  Many images are too small, or the part is not centered properly.  Also, one subModel has a different number of steps and step dats in it... need to track that down.  But getting closer to complete image generation...