Kill lookat vector in povray file fixup, since it's just not necessary.  With it, pov and opengl output doesn't line up.
Be sure to write the cached part size to the part, for future use.  Note that we're not caching CSI image size, so rendering CSI povs will fail - disable that code for now.

Move all path names and generation out of the individual renderPov functions in LDrawFileFormat, and into its global space.  Handy if we want to look up a filename in one of these directories from elsewhere.

Reformat some older code to be more consistent with new code (remove extraenous parentheses from if statements).