Add image size to CSI, both in the class itself and in the LDraw file line.  Use that number to generate povs.
One problem now remains: if a part or CSI was rendered partially off screen by opengl, opengl will displace the rendering to try and get it to fit on screen.  This displacement is added to the existing center displacement, which works great for opengl but not pov, because I can't get pov to displace the camera properly.  So, need to have GLHelpers.initImgSize return everything as it is PLUS the x & y displacement offset, then use that in CSI's (and PartOGL's) renderToPov.  To that by enlarging the pov image size by 2x the displacement in each x & y.

Once that's done, just need to include the rendered CSI in the instructions...