Fixup page image generation by first instantiating just one cairo context for the entire operation, then clearing it when necessary.  Much more efficient and easier to understand this way.  Win win.
Use new page numbering to create one "page_%d.png" image file for each page number.  Can now properly detect if we need to recreate a missing page image or not, which saves a ton of processing

Also, be sure to properly generate dats / povs / pngs for each part in a give step.  Do not rely on the PLIItem rendering that part, since PLIItem draw may not be called!  Very little work is done if the same part is rendered twice, but could check that just to make things even better.