Fix submodel page number drawing issue.  Problem was that we were drawing a given step's submodels in the middle of drawing that step!  Of course, since we were't finished drawing the current step, any drawing we did do was totally hosed by submodel drawing. Duh.  Fixed by splitting off current step drawing and current step's submodel drawing.
Also reorganized the overall image generation order.  Now, everything is drawn basically in the same order it appears in the gui tree, which is nice.  Also more straightforward this way, and less code.  Always nice.

Now, the only image generation issues remaining:
 - need to adust pov generation to take advantage of displaced camera
 - submodel image generation is all wrong, but that should be fixed by the previous point

That's it.  Should be done by end of day today... If nothing else explodes in spectacular flames of doom...