Huge progress!
First, fix annoying GL Buffer initialization problem, where the 4th step's csi in pyramid wouldn't show up.  This was because in the middle of initializing all the csi dimensions, I was jumping into each csi pixmap init, which hosed everything.  Duh.

But more importantly, the terrible evil crash-when-expanding-step3-pli is fixed.  Oh sweet victory.  The problem was stupid step.row(); a step's row has nothing to do with it's number; a step number is completely independent of the page its on, and is only relative to the previous step.  No, a steps row depends on where in the step array of the containing page it is.  Again, duh.

Also, clean up instruction.selectPage() so that it clears any existing selection when page is changed.  This very nicely handles multi-select across multiple pages, by simply destroying any existing selection.

Holy crap does pyqt friggin rock.  All this is so nice and clean compared to plugugly gtk.  Rad.

Also *finally* hit a second song on Black Halo, and it rocks.