Start work on exporting final rendered images through l3p and povray.  Didn't actually get near to either of those apps yet.  Still setting framework for being able to convert a CSI into a .DAT, given that we've totally ditched the imported LDraw file.
With this, there' some initial experimental export code in lic_qt.py's export menu handler; this correctly exports all non-pixmap items, which is wonderful.  Pixmap items should not be exported anyway, since those will be replaced by the high quality renderings from povray.

Moved the parts array from step to CSI, since step never used it at all.  It's where it belongs now.

With this, CSIs now point to their previous CSIs directly.  This link is created on model import, and is saved as a (page number, step number) pair for each CSI.  This works because the (page number, step number) pair uniquely identify each and every CSI inside an instruction book (ignoring callouts).

On load, that pair is saved to the csi.prevCSI as the tuple it is.  Load's last step is to walk the instruction book and convert any csi.prevCSI tuple to an actual csi pointer.

Seems to work just fine.  Will find out more when it comes time to re-render a csi that was loaded from a lic file, instead of relying on the saved pixmap...

Also hit a new DragonForce song - bloody finally!  It's insanely good too.