Start in on being able to select individual parts in a model, so they can be moved to the previous or next step, grouped into a submodel, whatever.
This adds each part as a tree element inside its parent PLIItem.  Still need to be able to draw a single part as selected or unselected.  This requires either some kind of open GL wireframe drawing, or some kind of alpha overlay.

Wireframe drawing would be a pain in the butt because all of the edge / line info in each part file is currently ignored.  We'd need to load all that info, save it, and keep it around everywhere.  May not be so bad since we carry only one copy of each part data, but still.

Alpha overlay could be pulled off by simply drawing the part a second time, color XORed, and semi-transparent, if open GL can do that...

Also, start in on a Page Size dialog.  But that's boring stuff, easy, and it's easy enough to just flip the hardcoded page variables in the code, for myself, for now.