Add support for moving steps to the previous / next page.  This includes code to re-layout any affected page.  For now, steps are ordered left to right, which is only correct if there's two steps.
This also introduces a children array to Page, which keeps all of its children.  Was getting way to confusing trying to guess what child was where.  This now makes adding things like a step separator or 1-to-1 axle image easier.

Also, fix up the CSIPLIResize dialog by creating it with the current CSI & PLI scale values.  And remember, when adding something to the undo stack, the 'redo' method *is called*.  This is important because if you do the action then add the action to the stack, you'll do the action twice.