Clean up progress bar, a little bit.  Actually, this muddles up the otherwise pretty clean code quite a lot.  Now, like 5 regular methods are suddenly generators, yielding crap everywhere.  But, the end result is rather nice - a clear, easy to use progress bar dialog, with fully supported cancel button.
Not sure if the code inside the Part size init is actually working or not - check that out.

An interesting test to run: try loading a big model both *with* and *without* all this generator and progress bar code - is load time significantly affected?  If so, is it worth it?