Begin the long work on callouts.  With this, callouts can be created from a selected part's context menu.  There's zero layout code yet; this focuses on getting the tree right.  Callouts sit inside their respective Step, with their own steps inside.  Steps in a Callout have their PLI disabled and correctly removed from the tree.  Biggest problem is what to do with individual parts - they now need to exist in original PLI, and in Callout's CSI, or somewhere...
Also, refactor BeginDisplacement undo command so it takes an arrow as an argument - saves importing Arrow from Model (a circular import).

Also, add a reference to the undo stack to the scene, which is available everywhere.  With this, it should be possible to remove all of the undo / redo signals everywhere, and just call stack.push directly when needed.  TODO: refactor this.

Also, change recent files list so that a newly opened file is bumped to the top of the list.