Add support for right clicking stuff in the tree.  This just brings up the identical context menu as would clicking in the graphics scene.  Nice.  Also clean up Scene's context menu, by using type instead of isinstance.  Nice.  This removes yet another non-obvious signal; move to the very straightforward mouseReleaseEvent instead.  This allows us to not do any of the selection work on right click.
And use the new item.getPage() method while we're at it, to clean up tree click and page scrolling.

Also, when a PartItem is clicked in the tree, force tree to scroll immediately to its first child; since PartItems are nothing but group holders for parts, and not selectable, this works well.