Some refactoring, to add support for (border/fill) formatting any template rect item that wants it.
Add new GraphicsRoundRectItem class, which extends QGraphicsRectItem by adding round corners, a corner radius, and a pen with corner attribute stuck to it.

Add new TemplateRectItem class, which is the first use of multiple inheritance in Lic so far.  This class encapsulates all menu / undo / redu functionality common to all graphics items that can format their border and fill.  To use, create a new class which subclasses both TemplateRectItem and the target Model class, like TemplateCallout(TemplateRectItem, Callout).  Can be an otherwise empty class.

Only problem with this is that each graphics item needs its own subclass, because in SetNewPen undo / redo, we compare the target's class to all children of each page, so as to apply the template.  Subclasses must (obviously) be compared.  But if an item subclasses QGraphicsRect directly, it too will be changed... So far this affects only submodelImage, which should be a full class anyway.  That's coming up.

Also, add a spin box to the border dialog that controls corner radius.