Add undo / redo support for adding & removing Title Page labels.  Add context menu option to Labels to delete them.
Fix new Layout code so it doesn't crash when positioning a row of only fixed-size items.

Fix Templates so they can be applied without crashing to Instruction books that have an empty PLI in the first Step.

Fix PLIs so that Submodel parts never ever get inserted into a PLI.